+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Features
+- program: add burn lp share cost to logging/record ([#814](https://github.com/drift-labs/protocol-v2/pull/814))
 
 ### Fixes